refactor(report): retire legacy report directory and migrate acceptance redaction
