+++ src/workflow/report-redact.ts
