+++ tests/legacy-ir-removal.test.ts
@@ -9,7 +9,7 @@ async function expectMissing(relativePath: string): Promise<void> {
 }
 
 describe('legacy IR to Playwright chain removal', () => {
-  it('removes the compiler, exploration, repair, classification, importer, schema, demo, and fixture files', async () => {
+  it('removes the compiler, exploration, repair, classification, importer, schema, integrated report, demo, and fixture files', async () => {
     const legacyPaths = [
       'src/compiler/playwright.ts',
       'src/importer.ts',
@@ -24,21 +24,29 @@ describe('legacy IR to Playwright chain removal', () => {
       'src/cli/validate-locators.ts',
       'src/cli/classify-failures.ts',
       'src/cli/repair.ts',
+      'src/cli/report.ts',
+      'src/report/playwright-json.ts',
+      'src/report/build.ts',
+      'src/report/html.ts',
+      'src/report/types.ts',
+      'src/report/redact.ts',
       'schemas/test-case-ir.schema.json',
       'examples/login-suite.ir.json',
       'examples/local-login-suite.ir.json',
       'playwright.demo.config.ts',
       'tests/fixtures/site/server.mjs',
+      'tests/report.test.ts',
     ]
     for (const path of legacyPaths) await expectMissing(path)
   })
 
-  it('removes the legacy npm commands while preserving replay compilation', async () => {
+  it('removes the legacy npm commands while preserving replay compilation and workflow acceptance reporting', async () => {
     const manifest = JSON.parse(await readFile(resolve(root, 'package.json'), 'utf8')) as { scripts: Record<string, string> }
-    for (const name of ['import', 'compile', 'explore', 'validate:locators', 'classify', 'repair', 'fixture', 'demo:compile', 'demo:test', 'demo:report']) {
+    for (const name of ['import', 'compile', 'explore', 'validate:locators', 'classify', 'repair', 'report', 'fixture', 'demo:compile', 'demo:test', 'demo:report']) {
       expect(manifest.scripts, name).not.toHaveProperty(name)
     }
     expect(manifest.scripts['compile:replay']).toBeDefined()
+    expect(manifest.scripts['report:workflow']).toBeDefined()
   })
 
   it('stops re-exporting the removed modules from the public index', async () => {
